requirements.txt
aniso8601==9.0.1
asttokens==2.0.8
backcall==0.2.0
catboost==1.0.5
category-encoders==2.5.0
certifi==2022.9.14
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
feather==0.1.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.37.3
graphviz==0.20.1
icecream==2.1.3
idna==2.10
importlib-metadata==4.12.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
kiwisolver==1.4.4
libzero==0.0.3.dev7
lightgbm==3.3.2
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numba==0.56.2
numpy==1.20.2
packaging==21.3
pandas==1.4.1
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.10.0
ply==3.11
progressbar2==4.0.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
Pygments==2.13.0
pynvml==8.0.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==5.0.2
python-utils==3.3.3
pytomlpp==1.0.11
pytz==2022.2.1
PyYAML==6.0
quantiphy==2.18.0
requests==2.25.1
scikit-learn==1.0.2
scipy==1.9.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
stack-data==0.5.0
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
xgboost==1.5.2
zipp==3.8.1
